+++ inventory/app/core/stock.py
@@ -1,5 +1,4 @@
 from dataclasses import dataclass
-from decimal import Decimal
 
 
 @dataclass
@@ -7,4 +6,4 @@ class Stock:
     id: int
     product_id: int
     warehouse_id: int
-    stock: Decimal
+    stock: float
